Split KafkaClusterTest into smaller test classes
